Update contact form block more the editor.js export to be an array
