+++ client/gutenberg/extensions/presets/jetpack/editor.js
@@ -8,16 +8,17 @@ import './editor-shared/block-category'; // Register the Jetpack category
 
 // TODO: Generate dyanmically from index.json
 // Appending `Block` to the names to keep `Map` from colliding with JS' Map
-import * as ContactFormBlock from 'gutenberg/extensions/contact-form/editor';
+import * as ContactFormBlock from 'gutenberg/extensions/contact-form';
 import * as MarkdownBlock from 'gutenberg/extensions/markdown';
 import * as MapBlock from 'gutenberg/extensions/map';
 import * as PublicizeBlock from 'gutenberg/extensions/publicize';
 import * as SimplePaymentsBlock from 'gutenberg/extensions/simple-payments';
 
-export default {
-	[ ContactFormBlock.name ]: ContactFormBlock.settings,
-	[ MarkdownBlock.name ]: MarkdownBlock.settings,
-	[ MapBlock.name ]: MapBlock.settings,
-	[ PublicizeBlock.name ]: PublicizeBlock.settings,
-	[ SimplePaymentsBlock.name ]: SimplePaymentsBlock.settings,
-};
+export default [
+	{ name: ContactFormBlock.name, settings: ContactFormBlock.settings },
+	...ContactFormBlock.fields,
+	MarkdownBlock,
+	MapBlock,
+	PublicizeBlock,
+	SimplePaymentsBlock,
+];
