+++ client/gutenberg/extensions/presets/jetpack/utils/refresh-registrations.js
@@ -2,7 +2,7 @@
 /**
  * External dependencies
  */
-import { forEach, get, has } from 'lodash';
+import { get, has } from 'lodash';
 import { getBlockType, registerBlockType, unregisterBlockType } from '@wordpress/blocks';
 import { getPlugin, registerPlugin, unregisterPlugin } from '@wordpress/plugins';
 
@@ -26,8 +26,8 @@ export default function refreshRegistrations() {
 	if ( ! extensionAvailability ) {
 		return;
 	}
-
-	forEach( extensions, ( settings, name ) => {
+	extensions.forEach( extension => {
+		const { name, settings } = extension;
 		const available = get( extensionAvailability, [ name, 'available' ] );
 
 		if ( has( settings, [ 'render' ] ) ) {
